pyproject.toml
[project]
name = "dashboard-commission"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.40.2",
    "gspread>=6.2.1",
    "jdatetime>=5.2.0",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "plotly>=6.1.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "streamlit>=1.45.1",
]
